+++ lib/roda/data_grid/search_grid_data.rb
@@ -13,6 +13,7 @@ module Crossbeams
         @client_rule_check = options.fetch(:client_rule_check)
         @config = SearchGridConfig.new(options)
         @params = parse_params(options)
+        @multi_dimensional_arrays = []
         assert_actions_ok!
       end
 
@@ -81,7 +82,7 @@ module Crossbeams
         in_keys
       end
 
-      def params_to_parms(params) # rubocop:disable Metrics/AbcSize
+      def params_to_parms(params) # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
         input_parameters = ::JSON.parse(params[:json_var]) || []
         parms = []
         # Check if this should become an IN parmeter (list of equal checks for a column.
@@ -199,6 +200,13 @@ module Crossbeams
           hs[:cellRenderer] = 'crossbeamsGridFormatters.iconFormatter' if col.name == 'icon'
           hs[:cellRenderer] = 'crossbeamsGridFormatters.barColourFormatter' if col.format == :bar_colour
 
+          # Sparkline chart formats
+          if SPARKTYPES.keys.include?(col.format)
+            hs[:cellRenderer] = 'agSparklineCellRenderer'
+            hs[:cellRendererParams] = { sparklineOptions: { type: SPARKTYPES[col.format] } }
+            @multi_dimensional_arrays << col.name.to_sym if col.format.to_s.end_with?('_text')
+          end
+
           # Rules for editable columns
           if edit_columns.include?(col.name)
             hs[:editable] = true
@@ -289,7 +297,7 @@ module Crossbeams
             # red_negative = styles.add_style num_fmt: 8
             delim4 = styles.add_style(format_code: '#,##0.0000;[Red]-#,##0.0000')
             delim2 = styles.add_style(format_code: '#,##0.00;[Red]-#,##0.00')
-            and_styles = { delimited_1000_4: delim4, delimited_1000: delim2 }
+            and_styles = { delimited_1000_4: delim4, delimited_1000: delim2 } # rubocop:disable Naming/VariableNumber
             report.ordered_columns.each do |col|
               xls_types << xls_possible_types[col.data_type] || :string # BOOLEAN == 0,1 ... need to change this to Y/N...or use format TRUE|FALSE...
               heads << col.caption
@@ -320,11 +328,11 @@ module Crossbeams
 
       private
 
-      def assert_actions_ok!
+      def assert_actions_ok! # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
         return unless config.actions
 
         config.actions.each do |action|
-          action.keys.each do |key|
+          action.each_key do |key|
             raise ArgumentError, "#{key} is not a valid action attribute" unless %i[
               auth
               has_permission
@@ -510,10 +518,11 @@ module Crossbeams
         DB[sql].map { |r| r.values.first }
       end
 
-      def dataminer_query(sql)
+      def dataminer_query(sql) # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity
         hstore = Object.const_defined?('Sequel::Postgres::HStore')
         DB[sql].to_a.map do |rec|
-          rec.keys.each do |key|
+          rec.each_key do |key|
+            rec[key] = rec[key].map { |a, b| [a, b.to_f] } if @multi_dimensional_arrays.include?(key)
             rec[key] = rec[key].to_f if rec[key].is_a?(BigDecimal)
             rec[key] = rec[key].to_s if hstore && rec[key].is_a?(Sequel::Postgres::HStore)
           end